initial component/page files
