+++ client/src/components/Footer/index.js
@@ -3,7 +3,7 @@ import React from 'react';
 const Footer = () => {
   return (
     <footer className="w-100 mt-auto bg-secondary p-4">
-      <div className="container">&copy;2020 by Lernantino</div>
+      <div className="container">&copy;2021 by Health Fusion</div>
     </footer>
   );
 };
