+++ client/src/components/Header/index.js
@@ -13,7 +13,7 @@ const Header = () => {
     <header className="bg-secondary mb-4 py-2 flex-row align-center">
       <div className="container flex-row justify-space-between-lg justify-center align-center">
         <Link to="/">
-          <h1>Deep Thoughts</h1>
+          <h1>Fusion-Storming</h1>
         </Link>
 
         <nav className="text-center">
