+++ client/src/components/ThoughtForm/index.js
@@ -1,11 +1,11 @@
 import React, { useState } from 'react';
 
 import { useMutation } from '@apollo/react-hooks';
-import { ADD_THOUGHT } from '../../utils/mutations';
-import { QUERY_THOUGHTS, QUERY_ME } from '../../utils/queries';
+import { ADD_DISCUSSION } from '../../utils/mutations';
+import { QUERY_DISCUSSION, QUERY_ME } from '../../utils/queries';
 
-const ThoughtForm = () => {
-  const [thoughtText, setText] = useState('');
+const DiscussionForm = () => {
+  const [discussionText, setText] = useState('');
   const [characterCount, setCharacterCount] = useState(0);
 
   const [addThought, { error }] = useMutation(ADD_THOUGHT, {
